Use declarative naming for Puppet classes
Puppet classes describe desired state. Class names should not use
action-verbs, such as "install" or "configure" in their names. Where
possible, class names should be declarative, and consist of nouns where
possible.

Parameters may be made available with adjective values to modify the
class definition.

The code in this manifest describes the configuration of the
trusted_external_command function in the context of the servicenow
integration. The desired state it describes (and applies) is the
servicenow_integration::trusted_external_command configuration. This
commit changes the class name to that.