Change snow_record.yaml to servicenow.yaml
User-facing terminology should use ServiceNow rather than the
abbreviation SNow. Plus, doing this neatly sidesteps the "-" vs. "_"
issue, at least for this file. Finally, built-in integration-related
files in the /etc/puppetlabs/puppet directory follow the convention of
being named after the component they provide integration configuration
for, so doing so for servicenow makes it fit in with the rest.

 - classifier.yaml
 - puppetdb.yaml