Prep get-servicenow-node-data.rb for PIE-255
This change will make the column to match the certname against
configurable because we anticipate that the name of the column that will
hold the data we need will either change soon, or will not be consistent
across customers.

It also formats the return string from the script so that there is a top
level `servicenow` json object, and all values are direct sub keys of
that object.

Most values returned to this query from ServiceNow will be their display
values and not their raw reference values. The exception seems to be
fields which appear to be complex object values that are not affected by
query parameter settings like `sysparm_display_value=true`

For example the field and value below are the same no matter the value
passed to the setting above.

```
"sys_domain": {
      "display_value": "global",
      "link": "https://dev84270.service-now.com/api/now/table/sys_user_group/global"
    }
```

This change also does additional value munging needed to make the
integration function properly.

- The required u_puppet* fields are munged in the script output to
  remove u_ portion of the field name
- The u_puppet_classes field is read, and the classes and variables
  found are parsed to create a proper "<class>::<variable>": "<value>"
  string.

Error checking has been added:
- If either the u_puppet_environment or the u_puppet_classes fields are
  missing, the script will throw an error.
- if the u_puppet_classes string is not valid json the script will throw
  an error.

If the script requests data for a node that does not exist in ServiceNow
the API will respond with simply `{"result": []}`. In this case the
script will now return `{"servicenow":{}}`.