(PIE-321) Backfill Gettings Started Guide to README
Backfilled missing portions of the getting started guide
(https://docs.google.com/document/d/1qcevkzvkXayNCZVrkENFTS51DKV4eA30ozU3hjZKOR4)
back to the README. We link to the getting started guide which currently
is a goolge doc, but will be replaced by the proper link once it's
published (PIE-322). Even though this README is a bit long I opted not
to break it into a docs folder because the relative links break the
forge README page unless you link to a specific release tag which takes
you back to github anyways. This is longer, but substantially simpler.