(PIE-385) Fix config validation dry run test
The module needs to be able to run a validation test on any parameters
that are stored in the configuration yaml file.

This change fixes a bug whereby it was possible to write a bad
configuration to disk and it was then impossible to fix it.

This new method of validation will be better at preventing an invalid
configuration from every being stored.