(PIE-606) Pin Litmus and Provision versions
This workaround pins the versions of the provision module and
puppet_litmus gem to work with the reporting_integration. Specifically
it restores functionality to provision with vmpooler and copy the
inventory file.

We also update the usage of three ServicenowTasks tasks to comply with
an updated version and get the tests running again.

Also updated ruby to 2.7.2 so that a requirement can be satisfied for
puppetfile-resolver.

Provision module is set to a specific SHA.
Puppet_litmus is set to version 0.26.0 or earlier (lowerbound of 0.18).